LPS-94013 Move related test class into SearchEngineUtilTest as inner class
