--- portal-impl/test/integration/com/liferay/portal/kernel/search/bundle/searchengineutil/TestSearchEngineConfigurator.java
@@ -1,49 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portal.kernel.search.bundle.searchengineutil;
-
-import com.liferay.portal.kernel.search.SearchEngine;
-import com.liferay.portal.kernel.search.SearchEngineConfigurator;
-
-import java.util.Map;
-import java.util.concurrent.atomic.AtomicBoolean;
-
-/**
- * @author Peter Fellwock
- */
-public class TestSearchEngineConfigurator implements SearchEngineConfigurator {
-
-	@Override
-	public void afterPropertiesSet() {
-		_atomicBoolean.set(Boolean.TRUE);
-	}
-
-	@Override
-	public void destroy() {
-		_atomicBoolean.set(Boolean.TRUE);
-	}
-
-	@Override
-	public void setSearchEngines(Map<String, SearchEngine> searchEngines) {
-		_atomicBoolean.set(Boolean.TRUE);
-	}
-
-	protected void setAtomicBoolean(AtomicBoolean atomicBoolean) {
-		_atomicBoolean = atomicBoolean;
-	}
-
-	private AtomicBoolean _atomicBoolean;
-
-}
